requirements.txt
#lista das dependências(libs). Ele vai instalar apartir desse arquivo usando o comando: pip install -r requirements.txt

Django==2.2.3
djangorestframework==3.10.2
Pillow==6.1.0
pkg-resources==0.0.0
pytz==2019.2
dj-database-url==0.5.0
sqlparse==0.3.0
gunicorn==19.9.0
whitenoise==4.1.3
psycopg2
